Task 3.3.3: Partly implement event processing
After wrestling with what looks like thread synchronization issues where in the second test events do not appear or come way late (after the testing thread shutdown), I decided to cut my losses and hand this in as it is.
Gauging from the LVA-forum on TUWEL it is clear that multiple students have similar issues and a problem with the testing code itself cannot be rule out, as there has not been any response from the team for more than 10 days (deadline <1week away, yay).